Refactoring using hooks
