+++ src/components/resultsContainer.tsx
@@ -1,7 +1,18 @@
 import * as React from "react";
 import { Result, Questions } from "../model/questionsResults";
+import { createStyles, WithStyles, withStyles } from "@material-ui/core";
+import DoneIcon from '@material-ui/icons/Done';
+import ClearIcon from '@material-ui/icons/Clear';
+const centeredElement = {"display": "flex", "align-items": "center", "justify-content": "center", "flex-direction": "column"};
 
-interface resultsContainerProps {
+const styles = theme =>
+  createStyles({
+    Score: centeredElement,
+    PlayAgain: {...centeredElement, "marginTop": "1rem"},
+    Result: {"display": "flex", "align-items": "center", "justify-content": "left", "border":"1px solid black", "padding": "0.5rem"}
+  });
+
+interface resultsContainerProps extends WithStyles<typeof styles> {
     result: Result;
     resetGame: () => void;
 }
@@ -10,32 +21,32 @@ const checkIfAnswerIsCorrect = (question: Questions): boolean => {
     return question.correct_answer.toLocaleLowerCase() === question.user_answer.toString();
 }
 
-const centeredElements = {"display": "flex", "align-items": "center", "justify-content": "center"};
-
-export const ResultsContainer = (props: resultsContainerProps) => {
-    const { result, resetGame } = props;
+const ResultsContainerComponent = (props: resultsContainerProps) => {
+    const { result, resetGame, classes } = props;
     const getScore = () => {
         return result.results.reduce((i,c,f) => checkIfAnswerIsCorrect(c) ? f++ : f, 0);
     }
     return (
         <React.Fragment>
-            <div style={centeredElements}>
-            <h3>You scored</h3>
-            <h3>{`${getScore()}/${result.results.length}`}</h3>
+            <div className={classes.Score}>
+            <h3>You scored </h3>
+            <h3>{` ${getScore()}/${result.results.length} `}</h3>
             </div>
             {
                 result.results.map(question => {
                     return (
-                        <div style={{"display": "flex", "align-items": "center", "justify-content": "left"}}>
-                            <span style={{"padding": 10}}>{checkIfAnswerIsCorrect(question) ? "+" : "-"}</span>
+                        <div className={classes.Result}>
+                            <span style={{"padding": 10}}>{checkIfAnswerIsCorrect(question) ? <DoneIcon/> : <ClearIcon/>}</span>
                             <p>{question.question}</p>
                         </div>
                     )
                 })
             }
-            {<div style={centeredElements}>
+            {<div className={classes.PlayAgain}>
             <button onClick={() => resetGame()}>Play Again?</button>
             </div>}
         </React.Fragment>
     );
 };
+
+export const ResultsContainer = withStyles(styles)(ResultsContainerComponent);
